refactor: use local category className
